slide-api/requirements.txt
celery>=5.2.3
fastapi>=0.70.0
uvicorn==0.17.5
pyvips==2.1.14
aiofiles==0.6.0
minio==7.1.0
python-multipart~=0.0.5
pymongo~=3.12.0
pydantic~=1.8.2
starlette>=0.16.0
watchdog>=2.1.6
PyYAML>=6.0
pydicom>=2.2.2
numpy>=1.21.5
matplotlib~=3.5.1
Pillow~=9.0.1
black>=22.3.0
loguru>=0.6.0